requirements2.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
blis==0.7.9
camelot-py==0.9.0
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
ci-info==0.3.0
click==8.1.3
cmake==3.26.3
confection==0.0.4
configobj==5.0.8
configparser==5.3.0
cryptography==40.0.2
cuda-python==12.1.0
cymem==2.0.7
Cython==0.29.34
decorator==5.1.1
Django==4.2.1
dnspython==2.3.0
easyocr==1.6.2
Elixir==0.7.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
etelemetry==0.3.0
filelock==3.12.0
fitz==0.0.1.dev2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
frozenlist==1.3.3
future==0.18.3
ghostscript==0.7
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
image==1.5.33
imageio==2.28.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
lazy_loader==0.2
lit==16.0.3
looseversion==1.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
networkx==3.1
nibabel==5.1.0
ninja==1.11.1
nipype==1.8.6
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.6
opencv-python-headless==4.5.4.60
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pathlib==1.0.1
pathy==0.10.1
pbr==5.11.1
pdfminer.six==20221105
Pillow==9.5.0
preshed==3.0.8
prov==2.0.0
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
PyJWT==2.7.0
pymongo==4.3.3
PyMuPDF==1.22.3
pyparsing==3.0.9
PyPDF2==2.12.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
pyxnat==1.5
PyYAML==6.0
rdflib==6.3.2
requests==2.30.0
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.4
srsly==2.4.6
sympy==1.12
Tempita==0.5.2
thinc==8.1.10
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
traits==6.3.2
triton==2.0.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
wasabi==1.1.1
Werkzeug==2.3.3
xlrd==0.7.1
xlwt==0.7.2
yarl==1.9.2
